+++ package.json
@@ -199,6 +199,7 @@
     "tslib": "2.6.3",
     "typescript": "5.5.4",
     "undici": "6.19.8",
+    "unenv": "^1.10.0",
     "verdaccio": "5.32.1",
     "verdaccio-auth-memory": "^10.0.0",
     "vite": "5.4.2",
