+++ packages/angular/ssr/BUILD.bazel
@@ -18,12 +18,12 @@ ts_library(
     ),
     module_name = "@angular/ssr",
     deps = [
+        "//packages/angular/ssr/third_party/critters:bundled_critters_lib",
         "@npm//@angular/common",
         "@npm//@angular/core",
         "@npm//@angular/platform-server",
         "@npm//@angular/router",
         "@npm//@types/node",
-        "@npm//critters",
     ],
 )
 
@@ -32,8 +32,15 @@ ng_package(
     package_name = "@angular/ssr",
     srcs = [
         ":package.json",
+        "//packages/angular/ssr/third_party/critters:bundled_critters_lib",
     ],
-    nested_packages = ["//packages/angular/ssr/schematics:npm_package"],
+    externals = [
+        "express",
+        "../../third_party/critters",
+    ],
+    nested_packages = [
+        "//packages/angular/ssr/schematics:npm_package",
+    ],
     tags = ["release-package"],
     deps = [
         ":ssr",
